style(components): move `ConnectButton` to `/auth`
